feat: fix login bug

+++ models/grouptour.js
@@ -23,7 +23,7 @@ module.exports = (sequelize, DataTypes) => {
     image: DataTypes.STRING,
     quantity: DataTypes.STRING,
     price: DataTypes.STRING,
-    rating: DataTypes.FLOAT,
+    rating: DataTypes.FLOAT(2, 1), // FLOAT(n, m)，n=總位數，m=小數位數
     canBeCancel: DataTypes.BOOLEAN,
     userId: DataTypes.INTEGER,
     categoryId: DataTypes.INTEGER
